Refactor networkingcommon: move subnets and spaces.
Networkingcommon/spaces moved to the spaces client facade.
Networkingcommon/subnets moved to the subnets client facade.

No need for this functionality in common, not used by more
than one facade.

BackingSpace, BackingSubnet and BackupSubnetInfo still used in
stub_network for tests, leave in networkingcommon until tests
can be rewritten.
