+++ apiserver/facades/client/spaces/export_test.go
@@ -3,4 +3,17 @@
 
 package spaces
 
+import (
+	"github.com/juju/juju/apiserver/common/networkingcommon"
+	"github.com/juju/juju/environs/context"
+)
+
 var NewAPIWithBacking = newAPIWithBacking
+
+func SupportsSpaces(backing networkingcommon.NetworkBacking, ctx context.ProviderCallContext) error {
+	api := &API{
+		backing: backing,
+		context: ctx,
+	}
+	return api.checkSupportsSpaces()
+}
