+++ apiserver/facades/client/spaces/shims.go
@@ -1,0 +1,61 @@
+// Copyright 2019 Canonical Ltd.
+// Licensed under the AGPLv3, see LICENCE file for details.
+
+package spaces
+
+import (
+	"github.com/juju/errors"
+	"gopkg.in/juju/names.v3"
+
+	"github.com/juju/juju/apiserver/common/networkingcommon"
+	"github.com/juju/juju/core/network"
+	"github.com/juju/juju/state"
+	"github.com/juju/juju/state/stateenvirons"
+)
+
+// NewStateShim returns a new state shim.
+func NewStateShim(st *state.State) (*stateShim, error) {
+	m, err := st.Model()
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	return &stateShim{EnvironConfigGetter: stateenvirons.EnvironConfigGetter{State: st, Model: m},
+		State: st, modelTag: m.ModelTag()}, nil
+}
+
+// stateShim forwards and adapts state.State methods to Backing
+// method.
+type stateShim struct {
+	stateenvirons.EnvironConfigGetter
+	*state.State
+	modelTag names.ModelTag
+}
+
+func (s *stateShim) ModelTag() names.ModelTag {
+	return s.modelTag
+}
+
+func (s *stateShim) AddSpace(name string, providerId network.Id, subnetIds []string, public bool) error {
+	_, err := s.State.AddSpace(name, providerId, subnetIds, public)
+	return err
+}
+
+func (s *stateShim) AllSpaces() ([]networkingcommon.BackingSpace, error) {
+	results, err := s.State.AllSpaces()
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	spaces := make([]networkingcommon.BackingSpace, len(results))
+	for i, result := range results {
+		spaces[i] = networkingcommon.NewSpaceShim(result)
+	}
+	return spaces, nil
+}
+
+func (s *stateShim) Subnet(cidr string) (networkingcommon.BackingSubnet, error) {
+	result, err := s.State.Subnet(cidr)
+	if err != nil {
+		return nil, errors.Trace(err)
+	}
+	return networkingcommon.NewSubnetShim(result), nil
+}
